[level 1] Title: 나누어 떨어지는 숫자 배열, Time: 0.41 ms, Memory: 75.9 MB -BaekjoonHub

+++ 프로그래머스/lv1/12910. 나누어 떨어지는 숫자 배열/나누어 떨어지는 숫자 배열.java
@@ -20,7 +20,7 @@ class Solution {
         
         // 리스트에 있는 값들을 배열에 담기
         for (int i = 0 ; i < list.size() ; i++) {
-            results[i] = list.get(i).intValue();
+            results[i] = list.get(i);
         }
         
         // 오름차순 정렬
